requirements.txt
django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
python-decouple
dj-config-url
celery[redis]
pillow
drf-yasg
django-filter
django-cors-headers
django-phonenumber-field[phonenumberslite]